change the folder structure
